bug fix: return only 1 prefix
On line 2403 we try to set the context node to the identifiable
element that has the `@ident` that matches the `@key` or `@name` of
the element tei:generateRefPrefix() as context, by using
`<xsl:for-each select="key('IDENTS',$lookup)">`. HOWEVER, it is
possible that there will be more than one element in IDENTS with the
same `@ident`. (E.g., this happens when the TEI-in-Libraries Best
Practices documentation is generated, because each of the 4 included
ODD files (level1.odd, level2.odd, level3.odd, level4.odd) has a
<specGrpRef> that points to a set of ODD elements in lib-header.odd.
This those particularly identifiable elements occur more than once in
IDENTS, and the `@ident` of one of them as a key would return 4 items.
This also might happen when elements of the same name occur in
different modules, which TEI itself does not do but another language
being defined in ODD might.) Thus the <xsl:for-each>, which I think
the original author intended to do nothing but set the context node,
ends up iterating multiple times, and we try to return multiple items,
when we are only allowed to return one.

To fix this, or at least to hack around it, I have changed the line in
question to read `<xsl:for-each select="key('IDENTS',$lookup)[1]">`,
thus selecting only the first element found in IDENTS with the
`@ident` of interest.

This may not be the ideal fix, but it is working for the case
intended, and has not (yet) broken anything else.